+++ components/IndexSectionIntroductionDescriptionCapybara.vue
@@ -1,0 +1,12 @@
+<script setup lang="ts">
+</script>
+
+<template>
+  <section>
+    <SectionTitle>水豚</SectionTitle>
+    <p>水豚是水豚屬下僅存的兩種生物之一。牠是一種半水棲的食草動物，也是世界上體型最大的齧齒類動物。原產於南美洲除了智利以外的所有稀樹草原和叢林中。</p>
+  </section>
+</template>
+
+<style scoped lang="sass">
+</style>
